Move diff logic to github
gitlab doesn't use it anymore.
