+++ internal/reporter/gitlab_logger_test.go
@@ -40,71 +40,6 @@ func TestGitlabLogger(t *testing.T) {
 	})
 }
 
-func TestDiffLineFor(t *testing.T) {
-	testCases := []struct {
-		name     string
-		lines    []diffLine
-		line     int64
-		expected diffLine
-		found    bool
-	}{
-		{
-			name:     "empty lines slice",
-			lines:    []diffLine{},
-			line:     5,
-			expected: diffLine{old: 0, new: 0, wasModified: false},
-			found:    false,
-		},
-		{
-			name: "exact match",
-			lines: []diffLine{
-				{old: 10, new: 10, wasModified: true},
-				{old: 11, new: 12, wasModified: false},
-			},
-			line:     10,
-			expected: diffLine{old: 10, new: 10, wasModified: true},
-			found:    true,
-		},
-		{
-			name: "line in gap - before first diff line",
-			lines: []diffLine{
-				{old: 10, new: 10, wasModified: true},
-			},
-			line:     5,
-			expected: diffLine{old: 5, new: 5, wasModified: false},
-			found:    true,
-		},
-		{
-			name: "line in gap - between diff lines",
-			lines: []diffLine{
-				{old: 5, new: 5, wasModified: true},
-				{old: 10, new: 12, wasModified: true},
-			},
-			line:     8,
-			expected: diffLine{old: 8, new: 8, wasModified: false},
-			found:    true,
-		},
-		{
-			name: "line after all diff lines",
-			lines: []diffLine{
-				{old: 5, new: 5, wasModified: true},
-				{old: 10, new: 10, wasModified: false},
-			},
-			line:     15,
-			expected: diffLine{old: 15, new: 15, wasModified: false},
-			found:    true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			result, found := diffLineFor(tc.lines, tc.line)
-			require.Equal(t, tc.found, found)
-			require.Equal(t, tc.expected, result)
-		})
-	}
-}
-
 func TestGetGitLabPaginated(t *testing.T) {
 	t.Run("single page", func(t *testing.T) {
 		items, err := getGitLabPaginated(func(_ int64) ([]string, *gitlab.Response, error) {
